+++ package.json
@@ -49,7 +49,7 @@
     "fork-ts-checker-webpack-plugin": "^4.1.3",
     "html-webpack-harddisk-plugin": "^1.0.1",
     "html-webpack-plugin": "^4.3.0",
-    "jest": "^25.5.3",
+    "jest": "^25.5.4",
     "jest-environment-jsdom-sixteen": "^1.0.3",
     "jest-junit": "^10.0.0",
     "lodash": "^4.17.15",
@@ -64,7 +64,7 @@
     "redux-mock-store": "^1.5.4",
     "resolve-url-loader": "^3.1.1",
     "sass-loader": "^8.0.2",
-    "terser-webpack-plugin": "^2.3.6",
+    "terser-webpack-plugin": "^3.0.0",
     "thread-loader": "^2.1.3",
     "ts-loader": "^7.0.2",
     "typescript": "^3.8.3",
@@ -95,7 +95,7 @@
     "install": "^0.13.0",
     "jsdom": "^16.2.2",
     "loaders.css": "^0.1.2",
-    "moment": "2.24.0",
+    "moment": "2.23.0",
     "moment-timezone": "^0.5.28",
     "numeral": "^2.0.6",
     "query-string": "^6.12.1",
