+++ www/front_src/src/Resources/Filter/storedFilter.ts
@@ -1,6 +1,6 @@
 import { isNil } from 'ramda';
 
-import { unhandledProblemsFilter, FilterGroup } from './Filter/models';
+import { unhandledProblemsFilter, FilterGroup } from './models';
 
 const key = 'centreon-events-filter';
